+++ PureInvoke/src/advapi32.cs
@@ -6,15 +6,15 @@ using System.Text;
 
 namespace PureInvoke
 {
-	public static class AdvApi32
+	public static class Advapi32
 	{
 		// ReSharper disable InconsistentNaming
-		[DllImport("advapi32", CharSet = CharSet.Auto, SetLastError = true)]
+		[DllImport("advapi32", CharSet=CharSet.Auto, SetLastError=true)]
 		public static extern bool ConvertSidToStringSid(
 			[MarshalAs(UnmanagedType.LPArray)] byte[] pSID,
 			out IntPtr ptrSid);
 
-		[DllImport("advapi32.dll", CharSet = CharSet.Auto, SetLastError = true)]
+		[DllImport("advapi32.dll", CharSet=CharSet.Auto, SetLastError=true)]
 		public static extern bool LookupAccountName(
 			string lpSystemName,
 			string lpAccountName,
@@ -25,7 +25,7 @@ namespace PureInvoke
 			out SidNameUse peUse
 		);
 
-		[DllImport("advapi32.dll", CharSet = CharSet.Auto, SetLastError = true)]
+		[DllImport("advapi32.dll", CharSet=CharSet.Auto, SetLastError=true)]
 		public static extern bool LookupAccountSid(
 			string lpSystemName,
 			[MarshalAs(UnmanagedType.LPArray)] byte[] Sid,
@@ -37,7 +37,7 @@ namespace PureInvoke
 		);
 
 		public enum SidNameUse {
-			User = 1,
+			User=1,
 			Group,
 			Domain,
 			Alias,