+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 
 * Created `Invoke-AdvapiLookupAccountName` to call the Advapi `LookupAccountName` function.
 * Created `Invoke-AdvapiLookupAccountSid` to call the Advapi `LookupAccountName` function.
+* Created `Invoke-KernelFindFileName` to call the Kernel `FindFirstFileNameW` and `FindNextFileNameW` functions to find
+all the hardlinks to a file.